pyproject.toml
[project]
name = "obsidian-rag"
version = "0.1.0"
description = "dependencies for my obsidian-rag work"
authors = [
    {name = "Ze Yuan Li",email = "zeyuan.bill.li@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.10,<4.0"
dependencies = [
    "sentence-transformers (>=3.3.1,<4.0.0)",
    "openai (>=1.59.8,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "numpy (<2.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "nicegui (>=2.10.0,<3.0.0)",
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
